% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_get_api.R
\name{api_error_body}
\alias{api_error_body}
\title{Extract error message from JSON body of an HTTP response}
\usage{
api_error_body(resp)
}
\arguments{
\item{resp}{An HTTP response object.}
}
\value{
A named character vector representing the flattened JSON content of the HTTP response body.
}
\description{
Extract error message from JSON body of an HTTP response
}
\examples{
\dontrun{
# Assuming an HTTP response object 'response' with a JSON error body
error_body <- ecopi_error_body(response)
}
}
