% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_choose_time_unit.R
\name{choose_time_unit}
\alias{choose_time_unit}
\title{Choose Time Unit}
\usage{
choose_time_unit(time_vector, num_units, possible_units)
}
\arguments{
\item{time_vector}{A vector containing time values.}

\item{num_units}{The desired number of units.}

\item{possible_units}{A named numeric vector containing possible time units to choose from.}
}
\value{
A character string indicating the name of the chosen time unit.
}
\description{
Automatically choose an appropriate time unit based on the range of values in a time vector.
}
\details{
This function calculates the length of the time interval in hours based on the range of values in the \code{time_vector} argument. It then calculates the ratio of the desired number of units to the length of the time interval for each possible time unit in the \code{possible_units} argument. The function selects the time unit that has the closest ratio to 1 and returns its name.

Note that the values in the \code{possible_units} vector should be in hours, and the names should correspond to valid time units that can be used with the \code{floor_date()} function in the \code{lubridate} package. If \code{lubridate} is not installed, only units in hours can be used.
}
\examples{
time_vector <- as.POSIXct(c("2022-02-01 12:00:00", "2022-02-15 18:00:00", "2022-03-01 00:00:00"))
possible_units <- c(`24 hours` = 24, `7 days` = 168, `1 month` = 720) # Corresponding values are in hours
choose_time_unit(time_vector, 2, possible_units)

}
\seealso{
\code{\link[base]{difftime}}, \code{\link[lubridate]{floor_date}}
}
\author{
Your Name Here
}
