% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_get_api.R
\name{get_log}
\alias{get_log}
\title{Retrieve log data from a given URL}
\usage{
get_log(url, params = NULL)
}
\arguments{
\item{url}{The URL of the API}

\item{params}{Parameters to GET request as list. Can include Django field lookups}
}
\value{
A dataframe containing log data.
}
\description{
Retrieve log data from a given URL
}
\examples{
params <- list(
  "recorder_nr_id" = "BirdNET-HI111",
  "datetime_pi__gte" = "2023-01-17"
)
get_log("https://reco.birdnet.tucmi.de/reco", params)

}
