% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_get_api.R
\name{get_recorders}
\alias{get_recorders}
\title{Retrieve recorder configs from a given URL}
\usage{
get_recorders(url, params = NULL)
}
\arguments{
\item{url}{The URL of the API}

\item{params}{Parameters to GET request as list. Can include Django field lookups}
}
\value{
Dataframe.
}
\description{
Retrieve recorder configs from a given URL
}
