% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_get_api.R
\name{patch_species_annotation}
\alias{patch_species_annotation}
\title{Update Species Annotation via a PATCH Request}
\usage{
patch_species_annotation(
  url,
  uid,
  confirmed = NULL,
  species_code_annotated = NULL
)
}
\arguments{
\item{url}{The base URL of the API endpoint.}

\item{uid}{A unique identifier for the species record to be updated.}

\item{confirmed}{A logical value indicating whether the annotation is confirmed. Default is NULL.}

\item{species_code_annotated}{A character string or NULL indicating the species code annotation. Default is NULL.}
}
\value{
The result of the performed request.
}
\description{
This function sends a PATCH request to update the species annotation for a given UID.
}
