% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_get_api.R
\name{perform_get_request}
\alias{perform_get_request}
\title{Perform get request}
\usage{
perform_get_request(url, path, params = NULL)
}
\arguments{
\item{url}{base URL to api}

\item{path}{path to append to base URL}

\item{params}{parameters to append to URL. Can include Django field lookups.}
}
\value{
API response
}
\description{
Perform get request
}
\examples{
perform_get_request("https://reco.birdnet.tucmi.de/reco", "det", list("recorder_id_id" = "BirdNET-HI111"))

}
