% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_get_api.R
\name{resp_body_json_to_df}
\alias{resp_body_json_to_df}
\title{Convert response body to data frame}
\usage{
resp_body_json_to_df(api_response)
}
\arguments{
\item{api_response}{Response from API}
}
\value{
Data frame
}
\description{
Convert response body to data frame
}
\examples{
api_response <- perform_get_request("https://reco.birdnet.tucmi.de/reco", "det", list("recorder_id_id" = "BirdNET-HI111"))
resp_body_json_to_df(api_response)
}
