% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_get_api.R
\name{send_json_request}
\alias{send_json_request}
\title{Send a JSON Request to a Given URL with the Specified Data and Resource Template}
\usage{
send_json_request(url, data, resource, ...)
}
\arguments{
\item{url}{The target URL to which the request will be sent.}

\item{data}{A list containing data to be sent in the request body in JSON format.}

\item{resource}{A character string specifying the resource template for the request.}

\item{...}{Additional arguments passed to the `req_template` function.}
}
\value{
The result of the performed request.
}
\description{
This function sends a JSON request to a specified URL, using a specified resource template.
The function constructs the request by adding user-agent, resource template, and request body.
}
\examples{
send_json_request("http://example.com/api/", list(name="John"), "POST /users/")
}
